# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Second Brain API",
    "description": "Knowledge management and learning system API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/knowledge/graph": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "Get Graph Visualization",
        "description": "Get graph data for visualization.\n\nReturns nodes and edges in D3-compatible format for force-directed graphs.\n\nArgs:\n    center_id: Optional node ID to center the graph on\n    node_types: Comma-separated list of node types to include\n    depth: How many hops from center to traverse (only used with center_id)\n    limit: Maximum number of nodes to return\n\nReturns:\n    GraphResponse with nodes, edges, and metadata",
        "operationId": "get_graph_visualization_api_knowledge_graph_get",
        "parameters": [
          {
            "name": "center_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Center graph on this node ID",
              "title": "Center Id"
            },
            "description": "Center graph on this node ID"
          },
          {
            "name": "node_types",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Comma-separated node types",
              "default": "Content,Concept,Note",
              "title": "Node Types"
            },
            "description": "Comma-separated node types"
          },
          {
            "name": "depth",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 4,
              "minimum": 1,
              "description": "Traversal depth from center",
              "default": 2,
              "title": "Depth"
            },
            "description": "Traversal depth from center"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 10,
              "description": "Max nodes to return",
              "default": 100,
              "title": "Limit"
            },
            "description": "Max nodes to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/knowledge/node/{node_id}": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "Get Node Details",
        "description": "Get detailed information about a specific node.\n\nArgs:\n    node_id: The node's unique identifier\n\nReturns:\n    NodeDetails with full node information\n\nRaises:\n    HTTPException 404: If node not found",
        "operationId": "get_node_details_api_knowledge_node__node_id__get",
        "parameters": [
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Node Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeDetails"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/knowledge/search": {
      "post": {
        "tags": [
          "knowledge"
        ],
        "summary": "Search Knowledge",
        "description": "Semantic search across knowledge base.\n\nUses embeddings to find semantically similar content when available,\notherwise falls back to keyword/fulltext matching. Searches across\nContent, Concepts, and Notes.\n\nArgs:\n    request: Search query and filters\n\nReturns:\n    SearchResponse with ranked results",
        "operationId": "search_knowledge_api_knowledge_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/knowledge/connections/{node_id}": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "Get Connections",
        "description": "Get connections for a specific node.\n\nReturns incoming and outgoing relationships, optionally filtered by direction.\n\nArgs:\n    node_id: The node to get connections for\n    direction: incoming, outgoing, or both\n    limit: Maximum connections per direction\n\nReturns:\n    ConnectionsResponse with incoming and outgoing lists",
        "operationId": "get_connections_api_knowledge_connections__node_id__get",
        "parameters": [
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Node Id"
            }
          },
          {
            "name": "direction",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ConnectionDirection",
              "description": "Filter by direction",
              "default": "both"
            },
            "description": "Filter by direction"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Max connections per direction",
              "default": 20,
              "title": "Limit"
            },
            "description": "Max connections per direction"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/knowledge/topics": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "Get Topic Hierarchy",
        "description": "Get hierarchical topic structure.\n\nTopics are organized in a tree based on their tag paths:\n- ml/\n  - ml/deep-learning/\n    - ml/deep-learning/transformers/\n\nArgs:\n    min_content: Filter out topics with fewer items\n\nReturns:\n    TopicHierarchyResponse with tree structure",
        "operationId": "get_topic_hierarchy_api_knowledge_topics_get",
        "parameters": [
          {
            "name": "min_content",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Min content count to include",
              "default": 0,
              "title": "Min Content"
            },
            "description": "Min content count to include"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopicHierarchyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConnectionDirection": {
        "type": "string",
        "enum": [
          "incoming",
          "outgoing",
          "both"
        ],
        "title": "ConnectionDirection",
        "description": "Direction filter for connection queries.\n\nUsed by the /connections/{node_id} endpoint to filter\nrelationships by direction relative to the queried node.\n\nValues:\n    INCOMING: Only relationships pointing to the node\n    OUTGOING: Only relationships pointing from the node\n    BOTH: All relationships (default)"
      },
      "ConnectionsResponse": {
        "properties": {
          "node_id": {
            "type": "string",
            "title": "Node Id"
          },
          "incoming": {
            "items": {
              "$ref": "#/components/schemas/NodeConnection"
            },
            "type": "array",
            "title": "Incoming"
          },
          "outgoing": {
            "items": {
              "$ref": "#/components/schemas/NodeConnection"
            },
            "type": "array",
            "title": "Outgoing"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "node_id",
          "incoming",
          "outgoing",
          "total"
        ],
        "title": "ConnectionsResponse",
        "description": "Response containing connections for a node.\n\nReturns both incoming and outgoing relationships,\nenabling bidirectional graph exploration.\n\nAttributes:\n    node_id: The queried node's ID\n    incoming: Connections pointing to this node\n    outgoing: Connections pointing from this node\n    total: Total connection count"
      },
...
      "GraphEdge": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "target": {
            "type": "string",
            "title": "Target"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "strength": {
            "type": "number",
            "maximum": 1,
            "minimum": 0,
            "title": "Strength",
            "default": 1
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          }
        },
        "type": "object",
        "required": [
          "source",
          "target",
          "type"
        ],
        "title": "GraphEdge",
        "description": "Lightweight edge for graph visualization.\n\nDesigned for bulk transfer in graph visualization responses where\nhundreds of edges may be returned. Contains only IDs and visual\nproperties - no display metadata to minimize payload size.\n\nFor detailed connection information with titles and context,\nsee NodeConnection which is used by the /connections endpoint.\n\nAttributes:\n    source: Source node ID\n    target: Target node ID\n    type: Relationship type (REFERENCES, RELATES_TO, etc.)\n    strength: Edge weight from 0.0 to 1.0 for visual thickness\n    label: Optional display label for the edge\n\nSee Also:\n    NodeConnection: Rich connection model with display metadata\n    GraphResponse: Uses GraphEdge for visualization data"
      },
      "GraphNode": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "content_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Type"
          },
          "size": {
            "type": "integer",
            "title": "Size",
            "description": "Node size weight based on connections",
            "default": 1
          },
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color",
            "description": "Optional hex color code"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata",
            "description": "Additional properties"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "type"
        ],
        "title": "GraphNode",
        "description": "Node in the knowledge graph visualization.\n\nRepresents a single entity (Content, Concept, or Note) in the graph\nwith properties needed for D3.js force-directed rendering.\n\nAttributes:\n    id: Unique node identifier (UUID or Neo4j ID)\n    label: Display title for the node\n    type: Node type - \"Content\", \"Concept\", or \"Note\"\n    content_type: For Content nodes, the specific type (paper, article, etc.)\n    size: Visual weight based on connection count (for node sizing)\n    color: Optional hex color code for custom styling\n    metadata: Additional properties (tags, etc.)"
      },
      "GraphResponse": {
        "properties": {
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/GraphNode"
            },
            "type": "array",
            "title": "Nodes"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/GraphEdge"
            },
            "type": "array",
            "title": "Edges"
          },
          "center_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Center Id"
          },
          "total_nodes": {
            "type": "integer",
            "title": "Total Nodes"
          },
          "total_edges": {
            "type": "integer",
            "title": "Total Edges"
          }
        },
        "type": "object",
        "required": [
          "nodes",
          "edges",
          "total_nodes",
          "total_edges"
        ],
        "title": "GraphResponse",
        "description": "Response containing graph data for visualization.\n\nReturns a subgraph suitable for rendering with D3.js or similar\nvisualization libraries. Includes metadata about the full graph size.\n\nAttributes:\n    nodes: List of nodes in the subgraph\n    edges: List of edges connecting the nodes\n    center_id: ID of the center node (if centered query)\n    total_nodes: Total nodes in the full graph (before limiting)\n    total_edges: Total edges in the returned subgraph"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NodeDetails": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "content_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Type"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "connections": {
            "type": "integer",
            "title": "Connections",
            "default": 0
          },
          "file_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Path"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "type"
        ],
        "title": "NodeDetails",
        "description": "Detailed information about a single node.\n\nReturns comprehensive metadata for a node detail view,\nincluding content summary, tags, and connection count.\n\nAttributes:\n    id: Node identifier\n    label: Display title\n    type: Node type (Content, Concept, Note)\n    content_type: Specific content type if applicable\n    summary: Brief summary text\n    tags: List of associated tags\n    source_url: Original source URL if applicable\n    created_at: Creation timestamp (ISO 8601)\n    connections: Count of connected nodes\n    file_path: Obsidian vault file path if applicable\n    name: Alternative name field (for Concepts)"
      },
...
      "SearchRequest": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Query"
          },
          "node_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Node Types",
            "default": [
              "Content",
              "Concept"
            ]
          },
          "limit": {
            "type": "integer",
            "maximum": 100,
            "minimum": 1,
            "title": "Limit",
            "default": 20
          },
          "min_score": {
            "type": "number",
            "maximum": 1,
            "minimum": 0,
            "title": "Min Score",
            "default": 0.5
          },
          "use_vector": {
            "type": "boolean",
            "title": "Use Vector",
            "description": "Use vector search when available",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "SearchRequest",
        "description": "Search query parameters.\n\nConfigures semantic search across the knowledge graph\nwith filtering and scoring options.\n\nAttributes:\n    query: Search query text (1-500 characters)\n    node_types: Types to search (default: Content, Concept)\n    limit: Maximum results to return (1-100, default: 20)\n    min_score: Minimum relevance score threshold (0-1, default: 0.5)\n    use_vector: Whether to use vector/embedding search when available"
      },
      "SearchResponse": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/SearchResult"
            },
            "type": "array",
            "title": "Results"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "search_time_ms": {
            "type": "number",
            "title": "Search Time Ms"
          }
        },
        "type": "object",
        "required": [
          "query",
          "results",
          "total",
          "search_time_ms"
        ],
        "title": "SearchResponse",
        "description": "Search results response.\n\nContains ranked search results with query echo\nand performance metrics.\n\nAttributes:\n    query: Original search query\n    results: List of matching results, ranked by score\n    total: Total number of results returned\n    search_time_ms: Search execution time in milliseconds"
      },
...
      "TopicHierarchyResponse": {
        "properties": {
          "roots": {
            "items": {
              "$ref": "#/components/schemas/TopicNode"
            },
            "type": "array",
            "title": "Roots"
          },
          "total_topics": {
            "type": "integer",
            "title": "Total Topics"
          },
          "max_depth": {
            "type": "integer",
            "title": "Max Depth"
          }
        },
        "type": "object",
        "required": [
          "roots",
          "total_topics",
          "max_depth"
        ],
        "title": "TopicHierarchyResponse",
        "description": "Full topic hierarchy response.\n\nReturns the complete topic tree for navigation\nwith aggregate statistics.\n\nAttributes:\n    roots: Root-level topic nodes\n    total_topics: Total number of topics in the tree\n    max_depth: Maximum depth of the tree"
      },
...
    }
  }
}